requirements.txt
apache-airflow-client @ git+https://github.com/apache/airflow-client-python.git@99b3d15620e543cf7b01da8f076bb78ae3eb9b71
celery==5.2.7
coverage==7.0.3
Flask==2.2.2
Flask-AppBuilder==4.1.6
Flask-Caching==2.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Flask-Migrate==4.0.4
flower==1.2.0
gunicorn==20.1.0
gviz-api==1.10.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
numpy==1.24.0
pandas==1.5.2
pandas-datareader==0.10.0
Pillow==8.4.0
plotly==5.11.0
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
PyYAML==6.0
redis==4.4.0
requests==2.28.1
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.39.0
tox==4.2.0
typing_extensions==4.4.0
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
Werkzeug==2.2.2
